+++ src/android/Badge.java
@@ -71,6 +71,9 @@ public class Badge extends CordovaPlugin {
         else if (action.equalsIgnoreCase("set")) {
             setBadge(args, callback);
         }
+        else if (action.equalsIgnoreCase("check")) {
+            checkSupport(callback);
+        }
         else {
             ret = false;
         }
@@ -88,7 +91,7 @@ public class Badge extends CordovaPlugin {
             @Override
             public void run() {
                 JSONObject cfg = impl.loadConfig();
-                sendPluginResult(callback, cfg);
+                callback.success(cfg);
             }
         });
     }
@@ -118,7 +121,7 @@ public class Badge extends CordovaPlugin {
             public void run() {
                 impl.clearBadge();
                 int badge = impl.getBadge();
-                sendPluginResult(callback, badge);
+                callback.success(badge);
             }
         });
     }
@@ -133,7 +136,7 @@ public class Badge extends CordovaPlugin {
             @Override
             public void run() {
                 int badge = impl.getBadge();
-                sendPluginResult(callback, badge);
+                callback.success(badge);
             }
         });
     }
@@ -153,31 +156,25 @@ public class Badge extends CordovaPlugin {
                 impl.clearBadge();
                 impl.setBadge(args.optInt(0));
                 int badge = impl.getBadge();
-                sendPluginResult(callback, badge);
+                callback.success(badge);
             }
         });
     }
 
     /**
-     * Send badge number has the plugin result back to the JS caller.
+     * Check support for badges.
      *
-     * @param callback The callback to invoke.
-     * @param badge    The badge number to pass with.
+     * @param callback The function to be exec as the callback.
      */
-    private void sendPluginResult(CallbackContext callback, int badge) {
-        PluginResult result = new PluginResult(OK, badge);
-        callback.sendPluginResult(result);
-    }
-
-    /**
-     * Send badge number has the plugin result back to the JS caller.
-     *
-     * @param callback The callback to invoke.
-     * @param obj      The object to pass with.
-     */
-    private void sendPluginResult(CallbackContext callback, JSONObject obj) {
-        PluginResult result = new PluginResult(OK, obj);
-        callback.sendPluginResult(result);
+    private void checkSupport (final CallbackContext callback) {
+        cordova.getThreadPool().execute(new Runnable() {
+            @Override
+            public void run() {
+                boolean support     = impl.isSupported();
+                PluginResult result = new PluginResult(OK, support);
+                callback.sendPluginResult(result);
+            }
+        });
     }
 
     /**